requirements.txt
acres
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
argcomplete
attrs
beautifulsoup4
blinker
cachetools
certifi
chardet
charset-normalizer
ci-info
click
colorama
compressed-rtf
configobj
configparser
dataclasses-json
defusedxml
distro
docx2txt
ebcdic
etelemetry
extract-msg
faiss-cpu
filelock
frozenlist
gitdb
GitPython
greenlet
h11
httpcore
httplib2
httpx
httpx-sse
idna
IMAPClient
importlib_resources
isodate
Jinja2
jiter
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
looseversion
lxml
markdown-it-py
MarkupSafe
marshmallow
mdurl
multidict
mypy-extensions
narwhals
networkx
nibabel
nipype
numpy
odfpy
olefile
openai
orjson
packaging
pandas
pathlib
pdfminer.six
pillow
propcache
protobuf
prov
puremagic
pyarrow
pycryptodome
pydantic
pydantic-settings
pydantic_core
pydeck
pydot
Pygments
PyMuPDF
pyparsing
python-dateutil
python-docx
python-dotenv
python-pptx
pytz
pyxnat
PyYAML
rdflib
referencing
regex
requests
requests-toolbelt
rich
rpds-py
scipy
simplejson
six
smmap
sniffio
sortedcontainers
soupsieve
SpeechRecognition
SQLAlchemy
streamlit
tenacity
textract
tiktoken
toml
tornado
tqdm
traits
typing-inspect
typing_extensions
tzdata
tzlocal
urllib3
watchdog
xlrd
XlsxWriter
yarl
zstandard
